refactor: Controller에서 사용하는 DTO의 패키지 경로 변경
