+++ src/main/java/roomescape/controller/dto/ReservationTimeCreateRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 public record ReservationTimeCreateRequest(
         String startAt
