+++ src/main/java/roomescape/controller/mapper/ReservationTimeMapper.java
@@ -1,7 +1,7 @@
 package roomescape.controller.mapper;
 
 import org.springframework.stereotype.Component;
-import roomescape.dto.ReservationTimeCreateRequest;
+import roomescape.controller.dto.ReservationTimeCreateRequest;
 import roomescape.service.command.ReservationTimeCreateCommand;
 
 @Component
